api/requirements.txt
Flask
google-cloud-storage
google-cloud-firestore
google-cloud-vision
python-dotenv
gunicorn
Flask-API